requirements.txt
aio-pika==9.3.1
aiocouch==3.0.0
aiohttp==3.9.1
aiokafka==0.10.0
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.1.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
Deprecated==1.2.14
fast-depends==2.2.8
faststream==0.3.11
frozenlist==1.4.1
google-api-core==2.15.0
google-auth==2.26.0
google-cloud-bigquery==3.14.1
google-cloud-core==2.4.1
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
iniconfig==2.0.0
multidict==6.0.4
packaging==23.2
pamqp==3.2.1
pluggy==1.3.0
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
pytest==7.4.4
pytest-asyncio==0.23.3
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
typer==0.9.0
typing_extensions==4.9.0
urllib3==2.1.0
wrapt==1.16.0
yarl==1.9.4
